pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]

[project]
name        = "sentseg-vihsd"
version     = "0.3.0"
description = "Vietnamese sentence segmentation on VIHSD"
authors     = [{name = "Research Upgraded"}]

dependencies = [
  "pandas>=2.2.2",
  "sklearn-crfsuite>=0.3.6",
  "python-crfsuite>=0.9.9",
  "pyyaml>=6.0",
  "nltk>=3.8.1",
  "tqdm"
]
